Moved executorContext builder into tester
