+++ core/src/main/scala/com/yahoo/maha/utils/GetTotalRowsRequest.scala
@@ -12,44 +12,6 @@ import scala.util.Try
 
 object GetTotalRowsRequest extends Logging {
 
-  def updateRowList(rowList: RowList) : Unit = {
-    val row = rowList.newRow
-    rowList.columnNames.foreach {
-      col =>
-        row.addValue(col, s"$col-value")
-    }
-    rowList.addRow(row)
-  }
-
-  def getQueryExecutorContext: QueryExecutorContext = {
-    val qeOracle = new QueryExecutor {
-      override def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
-        updateRowList(rowList)
-        QueryResult(rowList, queryAttributes, QueryResultStatus.SUCCESS)
-      }
-      override def engine: Engine = OracleEngine
-    }
-
-    val qeHive = new QueryExecutor {
-      override def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
-        updateRowList(rowList)
-        QueryResult(rowList, queryAttributes, QueryResultStatus.SUCCESS)      }
-      override def engine: Engine = HiveEngine
-    }
-    val qeDruid = new QueryExecutor {
-      override def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
-        updateRowList(rowList)
-        QueryResult(rowList, queryAttributes, QueryResultStatus.SUCCESS)      }
-      override def engine: Engine = DruidEngine
-    }
-
-    val qec = new QueryExecutorContext
-    qec.register(qeOracle)
-    qec.register(qeHive)
-    qec.register(qeDruid)
-    qec
-  }
-
   def getTotalRowsRequest(request: ReportingRequest, pipeline: QueryPipeline) : Try[ReportingRequest] = {
     //no filters except fk filters
     Try {
@@ -73,13 +35,12 @@ object GetTotalRowsRequest extends Logging {
     }
   }
 
-  def getTotalRows(request: RequestModel, sourcePipeline: QueryPipeline, registry: Registry)(implicit queryGeneratorRegistry: QueryGeneratorRegistry) : Try[Int] = {
+  def getTotalRows(request: RequestModel, sourcePipeline: QueryPipeline, registry: Registry, queryContext: QueryExecutorContext)(implicit queryGeneratorRegistry: QueryGeneratorRegistry) : Try[Int] = {
     Try {
       val totalRowsRequest: ReportingRequest = getTotalRowsRequest(request.reportingRequest, sourcePipeline).get
       val model: RequestModel = RequestModel.from(totalRowsRequest, registry).get
       val maxRows: Int = DruidQueryGenerator.defaultMaximumMaxRows
       assert(model.maxRows <= maxRows, throw new Exception(s"Value of ${model.maxRows} exceeds posted limit of $maxRows"))
-      val queryContext: QueryExecutorContext = getQueryExecutorContext
 
       val queryPipelineFactory = new DefaultQueryPipelineFactory()
