+++ core/src/test/scala/com/yahoo/maha/utils/GetTotalRowsRequestTest.scala
@@ -3,12 +3,50 @@ package com.yahoo.maha.utils
 import com.yahoo.maha.core.query.druid.DruidQueryGenerator
 import com.yahoo.maha.core.query.hive.HiveQueryGenerator
 import com.yahoo.maha.core.query.oracle.OracleQueryGenerator
-import com.yahoo.maha.core.{DefaultPartitionColumnRenderer, RequestModel, TestUDFRegistrationFactory}
-import com.yahoo.maha.core.query.{BaseQueryContextTest, BaseQueryGeneratorTest, SharedDimSchema}
+import com.yahoo.maha.core.{DefaultPartitionColumnRenderer, DruidEngine, Engine, HiveEngine, OracleEngine, RequestModel, TestUDFRegistrationFactory}
+import com.yahoo.maha.core.query._
 import com.yahoo.maha.core.request.ReportingRequest
 import org.scalatest.{BeforeAndAfterAll, FunSuite, Matchers}
 
 class GetTotalRowsRequestTest extends FunSuite with Matchers with BeforeAndAfterAll with BaseQueryGeneratorTest with SharedDimSchema with BaseQueryContextTest {
+
+  def updateRowList(rowList: RowList) : Unit = {
+    val row = rowList.newRow
+    rowList.columnNames.foreach {
+      col =>
+        row.addValue(col, s"$col-value")
+    }
+    rowList.addRow(row)
+  }
+
+  def getQueryExecutorContext: QueryExecutorContext = {
+    val qeOracle = new QueryExecutor {
+      override def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
+        updateRowList(rowList)
+        QueryResult(rowList, queryAttributes, QueryResultStatus.SUCCESS)
+      }
+      override def engine: Engine = OracleEngine
+    }
+
+    val qeHive = new QueryExecutor {
+      override def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
+        updateRowList(rowList)
+        QueryResult(rowList, queryAttributes, QueryResultStatus.SUCCESS)      }
+      override def engine: Engine = HiveEngine
+    }
+    val qeDruid = new QueryExecutor {
+      override def execute[T <: RowList](query: Query, rowList: T, queryAttributes: QueryAttributes) : QueryResult[T] = {
+        updateRowList(rowList)
+        QueryResult(rowList, queryAttributes, QueryResultStatus.SUCCESS)      }
+      override def engine: Engine = DruidEngine
+    }
+
+    val qec = new QueryExecutorContext
+    qec.register(qeOracle)
+    qec.register(qeHive)
+    qec.register(qeDruid)
+    qec
+  }
 
   override protected def beforeAll(): Unit = {
     OracleQueryGenerator.register(queryGeneratorRegistry, DefaultPartitionColumnRenderer)
@@ -72,7 +110,9 @@ class GetTotalRowsRequestTest extends FunSuite with Matchers with BeforeAndAfter
     assert(queryPipelineTry.isSuccess, queryPipelineTry.errorMessage("Fail to get the query pipeline"))
     val pipeline = queryPipelineTry.toOption.get
 
-    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry)
+    val queryContext: QueryExecutorContext = getQueryExecutorContext
+
+    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry, queryContext)
     assert(totalRowRequest.isSuccess, "Total Row Request Failed!")
   }
 
@@ -86,7 +126,9 @@ class GetTotalRowsRequestTest extends FunSuite with Matchers with BeforeAndAfter
     assert(queryPipelineTry.isSuccess, queryPipelineTry.errorMessage("Fail to get the query pipeline"))
     val pipeline = queryPipelineTry.toOption.get
 
-    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry)
+    val queryContext: QueryExecutorContext = getQueryExecutorContext
+
+    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry, queryContext)
     assert(totalRowRequest.isSuccess, "Total Row Request Failed!")
   }
 
@@ -100,7 +142,9 @@ class GetTotalRowsRequestTest extends FunSuite with Matchers with BeforeAndAfter
     assert(queryPipelineTry.isSuccess, queryPipelineTry.errorMessage("Fail to get the query pipeline"))
     val pipeline = queryPipelineTry.toOption.get
 
-    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry)
+    val queryContext: QueryExecutorContext = getQueryExecutorContext
+
+    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry, queryContext)
     assert(totalRowRequest.isSuccess, "Requests outside of Oracle should trace back into Oracle, if applicable")
   }
 
@@ -114,7 +158,9 @@ class GetTotalRowsRequestTest extends FunSuite with Matchers with BeforeAndAfter
     assert(queryPipelineTry.isSuccess, queryPipelineTry.errorMessage("Fail to get the query pipeline"))
     val pipeline = queryPipelineTry.toOption.get
 
-    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry)
+    val queryContext: QueryExecutorContext = getQueryExecutorContext
+
+    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry, queryContext)
     assert(totalRowRequest.isSuccess, "Requests outside of Oracle should trace back into Oracle, if applicable")
   }
 
@@ -128,7 +174,9 @@ class GetTotalRowsRequestTest extends FunSuite with Matchers with BeforeAndAfter
     assert(queryPipelineTry.isSuccess, queryPipelineTry.errorMessage("Fail to get the query pipeline"))
     val pipeline = queryPipelineTry.toOption.get
 
-    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry)
+    val queryContext: QueryExecutorContext = getQueryExecutorContext
+
+    val totalRowRequest = GetTotalRowsRequest.getTotalRows(requestModel.get, pipeline, registry, queryContext)
     assert(!totalRowRequest.isSuccess, "requested total tows should be less than the default max.")
   }
 }
